pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "anubis"
version = "0.1.0"
description = "AI-powered Windows PC optimization assistant using local LLMs and multi-agent orchestration"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Anubis Contributors" },
]
keywords = ["windows", "optimization", "llm", "agent", "system-monitoring"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    # Agent orchestration
    "langgraph>=0.2",
    "langchain-core>=0.3",
    "langchain-ollama>=0.2",

    # System monitoring
    "psutil>=6.0",
    "wmi>=1.5; sys_platform == 'win32'",
    "pywin32>=306; sys_platform == 'win32'",
    "pySMART>=1.3",

    # LLM integration
    "ollama>=0.4",
    "httpx>=0.27",

    # Data & storage
    "pydantic>=2.7",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",

    # API
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",

    # CLI
    "rich>=13.7",
    "typer>=0.12",

    # Utilities
    "pyyaml>=6.0",
    "structlog>=24.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "mypy>=1.10",
]
hardware = [
    # Requires LibreHardwareMonitor DLLs
    "pythonnet>=3.0",
]

[project.scripts]
anubis = "anubis.cli.main:app"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
